Test redirection

+++ app/controllers/concerns/current_analysis.rb
@@ -8,13 +8,15 @@ module CurrentAnalysis
 
   def current_analysis
     if session[:analysis_id].present?
-        begin
-          Analysis::Record.find_by(id: session[:analysis_id])
-        rescue ActiveRecord::RecordNotFound
+      Analysis::Record.find_by(id: session[:analysis_id]).then do |record|
+        if record.nil?
           Rails.logger.fatal "Analysis #{session[:analysis_id]} not found, this is a bug"
           session.delete(:analysis_id)
           current_analysis
+        else
+          record
         end
+      end
     else
       record = Analysis::Record.create(status: :pending)
       session[:analysis_id] = record.id
